feat!: add prefix digit for IDs
This is a hack to make the traversal order sorted.  The number-like keys will be traversed first in ascending order by value, adding prefix digit helps IDs starting with zero to be treated as numbers.

Another way is using arrays like `[id, data]`, and `Map` can be used to keep the order and allow random access at the same time.